+++ src/utils/findProject.ts
@@ -1,11 +1,7 @@
 
 import * as vscode from 'vscode';
+import * as path from 'path';
 
-
-// export async function findProjectFile(): Promise<vscode.Uri | undefined> {
-//   const files = await vscode.workspace.findFiles('**/*.xsproj', '**/node_modules/**', 1);
-//   return files[0];
-// }
 export async function findProjectFile(): Promise<vscode.Uri | null> {
   const files = await vscode.workspace.findFiles(
     '**/*.xsproj',
@@ -18,4 +14,15 @@ export async function findProjectFile(): Promise<vscode.Uri | null> {
   }
 
   return files[0];
+}
+
+export async function prepareProjectCwd(): Promise<string | null> {
+  if (!vscode.workspace.workspaceFolders) {
+    vscode.window.showErrorMessage('No Folder open.');
+    return null;
+  }
+  await vscode.workspace.saveAll();
+  const projectFile = await findProjectFile();
+  if (!projectFile) return null;
+  return path.dirname(projectFile.fsPath);
 }